Adding subcomponents for marketing content section
