+++ src/views/product-root-docs-path-landing/components/index.ts
@@ -1,9 +1,11 @@
 import ProductRootDocsPathLandingHero from './hero'
 import ProductRootDocsPathLandingIconCardLinkGrid from './icon-card-link-grid'
+import ProductRootDocsPathLandingMarketingContent from './marketing-content'
 import ProductRootDocsPathLandingPageHeading from './page-heading'
 
 export {
   ProductRootDocsPathLandingHero,
   ProductRootDocsPathLandingIconCardLinkGrid,
+  ProductRootDocsPathLandingMarketingContent,
   ProductRootDocsPathLandingPageHeading,
 }
